refactoring code
Had to use MMA options as several of them.
This facenet uses convolution all the way through without any concept of
padding. The old version hacked a non fully convolutional net (ie with
padding) and then filtered it back through the original trained net as a
check, so they are not exactly the same. There is some dispute as to
exactly how useful padding is, but for a convolutional system would
require an extra filtering stage.